pyproject.toml
[tool.poetry]
name = "auth_package"
version = "0.1.0"
description = "Generic robust authentication system"
authors = ["A K M Muhibullah Nayem <nayem.drmc@gmail.com>"]
license = "MIT"
repository = "https://github.com/MuhibNayem/auth-manager.git"

include = ["auth_package/examples/**"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.100.0"
sqlalchemy = "^2.0"
passlib = "^1.7.4"
aioredis = "^2.0.1"
motor = "^3.0.0"
boto3 = "^1.28.0"
botocore = "^1.31.0"
asyncpg = "^0.29.0"
uvicorn = "^0.22.0"
pyjwt = "^2.9.0"
requests = "^2.32.3"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
pyotp = "^2.9.0"
mailjet-rest = "^1.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
